requirements.txt
fastapi
uvicorn
transformers
accelerate
python-dotenv
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121